requirements.txt
fastapi
uvicorn[standard]
python-dotenv
email-validator
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
pymongo
openai
